refactor: reorganize utils structure
