+++ app/src/main/java/com/example/partymaker/PartyApplication.java
@@ -7,8 +7,8 @@ import com.example.partymaker.data.api.FirebaseServerClient;
 import com.example.partymaker.data.api.NetworkManager;
 import com.example.partymaker.data.firebase.DBRef;
 import com.example.partymaker.data.repository.GroupRepository;
-import com.example.partymaker.utils.MemoryManager;
-import com.example.partymaker.utilities.NotificationHelper;
+import com.example.partymaker.utils.system.MemoryManager;
+import com.example.partymaker.utils.ui.NotificationHelper;
 import com.google.firebase.FirebaseApp;
 
 /** Application class for PartyMaker. Initializes repositories and other app-wide components. */
