+++ app/src/main/java/com/example/partymaker/ui/group/JoinGroupActivity.java
@@ -1,7 +1,7 @@
 package com.example.partymaker.ui.group;
 
-import static com.example.partymaker.utilities.Common.hideViews;
-import static com.example.partymaker.utilities.Common.showViews;
+import static com.example.partymaker.utils.ui.Common.hideViews;
+import static com.example.partymaker.utils.ui.Common.showViews;
 
 import android.annotation.SuppressLint;
 import android.content.Intent;
@@ -18,10 +18,10 @@ import androidx.cardview.widget.CardView;
 import com.example.partymaker.R;
 import com.example.partymaker.data.api.FirebaseServerClient;
 import com.example.partymaker.data.model.Group;
-import com.example.partymaker.utilities.AuthHelper;
-import com.example.partymaker.utilities.Common;
-import com.example.partymaker.utilities.ExtrasMetadata;
-import com.example.partymaker.utilities.MapUtilities;
+import com.example.partymaker.utils.auth.AuthHelper;
+import com.example.partymaker.utils.ui.Common;
+import com.example.partymaker.utils.ui.ExtrasMetadata;
+import com.example.partymaker.utils.ui.MapUtilities;
 import java.util.HashMap;
 import java.util.Objects;
