+++ app/src/main/java/com/example/partymaker/ui/group/PartyMainActivity.java
@@ -1,7 +1,7 @@
 package com.example.partymaker.ui.group;
 
-import static com.example.partymaker.utilities.Common.hideViews;
-import static com.example.partymaker.utilities.Common.showViews;
+import static com.example.partymaker.utils.ui.Common.hideViews;
+import static com.example.partymaker.utils.ui.Common.showViews;
 
 import android.animation.ObjectAnimator;
 import android.annotation.SuppressLint;
@@ -28,11 +28,11 @@ import com.example.partymaker.R;
 import com.example.partymaker.data.api.FirebaseServerClient;
 import com.example.partymaker.data.model.ChatMessage;
 import com.example.partymaker.data.model.Group;
-import com.example.partymaker.utilities.AuthHelper;
-import com.example.partymaker.utilities.Common;
-import com.example.partymaker.utilities.ExtrasMetadata;
-import com.example.partymaker.utilities.NotificationHelper;
-import com.example.partymaker.utilities.ShareHelper;
+import com.example.partymaker.utils.auth.AuthHelper;
+import com.example.partymaker.utils.ui.Common;
+import com.example.partymaker.utils.ui.ExtrasMetadata;
+import com.example.partymaker.utils.ui.NotificationHelper;
+import com.example.partymaker.utils.sharing.ShareHelper;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 import java.util.HashMap;
