+++ app/src/main/java/com/example/partymaker/ui/group/UsersListActivity.java
@@ -17,9 +17,9 @@ import com.example.partymaker.data.firebase.FirebaseAccessManager;
 import com.example.partymaker.data.model.Group;
 import com.example.partymaker.data.model.User;
 import com.example.partymaker.ui.adapters.UserAdapter;
-import com.example.partymaker.utilities.AuthHelper;
-import com.example.partymaker.utilities.Common;
-import com.example.partymaker.utilities.ExtrasMetadata;
+import com.example.partymaker.utils.auth.AuthHelper;
+import com.example.partymaker.utils.ui.Common;
+import com.example.partymaker.utils.ui.ExtrasMetadata;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
