+++ app/src/main/java/com/example/partymaker/ui/profile/EditProfileActivity.java
@@ -30,8 +30,8 @@ import com.example.partymaker.data.firebase.DBRef;
 import com.example.partymaker.data.model.User;
 import com.example.partymaker.ui.auth.LoginActivity;
 import com.example.partymaker.ui.settings.ServerSettingsActivity;
-import com.example.partymaker.utilities.AuthHelper;
-import com.example.partymaker.utilities.BottomNavigationHelper;
+import com.example.partymaker.utils.auth.AuthHelper;
+import com.example.partymaker.utils.navigation.BottomNavigationHelper;
 import com.example.partymaker.viewmodel.UserViewModel;
 import com.google.android.material.snackbar.Snackbar;
 import com.google.firebase.storage.FirebaseStorage;
