+++ app/src/main/java/com/example/partymaker/utils/media/GlideImageLoader.java
@@ -1,4 +1,4 @@
-package com.example.partymaker.utils;
+package com.example.partymaker.utils.media;
 
 import android.content.Context;
 import android.graphics.drawable.Drawable;
