+++ app/src/main/java/com/example/partymaker/utils/navigation/BottomNavigationHelper.java
@@ -1,4 +1,4 @@
-package com.example.partymaker.utilities;
+package com.example.partymaker.utils.navigation;
 
 import android.app.Activity;
 import android.content.Intent;
