+++ app/src/main/java/com/example/partymaker/utils/server/ServerModeHelper.java
@@ -1,4 +1,4 @@
-package com.example.partymaker.utilities;
+package com.example.partymaker.utils.server;
 
 import android.content.Context;
 import android.content.SharedPreferences;
