+++ app/src/main/java/com/example/partymaker/utils/ui/Common.java
@@ -1,21 +1,21 @@
-package com.example.partymaker.utilities;
+package com.example.partymaker.utils.ui;
 
-import static com.example.partymaker.utilities.Constants.ADMIN_KEY;
-import static com.example.partymaker.utilities.Constants.CAN_ADD;
-import static com.example.partymaker.utilities.Constants.COMING_KEYS;
-import static com.example.partymaker.utilities.Constants.CREATED_AT;
-import static com.example.partymaker.utilities.Constants.DEFAULT_KEY;
-import static com.example.partymaker.utilities.Constants.FRIEND_KEYS;
-import static com.example.partymaker.utilities.Constants.GROUP_DAYS;
-import static com.example.partymaker.utilities.Constants.GROUP_HOURS;
-import static com.example.partymaker.utilities.Constants.GROUP_KEY;
-import static com.example.partymaker.utilities.Constants.GROUP_LOCATION;
-import static com.example.partymaker.utilities.Constants.GROUP_MONTHS;
-import static com.example.partymaker.utilities.Constants.GROUP_NAME;
-import static com.example.partymaker.utilities.Constants.GROUP_PRICE;
-import static com.example.partymaker.utilities.Constants.GROUP_TYPE;
-import static com.example.partymaker.utilities.Constants.GROUP_YEARS;
-import static com.example.partymaker.utilities.Constants.MESSAGE_KEYS;
+import static com.example.partymaker.utils.Constants.ADMIN_KEY;
+import static com.example.partymaker.utils.Constants.CAN_ADD;
+import static com.example.partymaker.utils.Constants.COMING_KEYS;
+import static com.example.partymaker.utils.Constants.CREATED_AT;
+import static com.example.partymaker.utils.Constants.DEFAULT_KEY;
+import static com.example.partymaker.utils.Constants.FRIEND_KEYS;
+import static com.example.partymaker.utils.Constants.GROUP_DAYS;
+import static com.example.partymaker.utils.Constants.GROUP_HOURS;
+import static com.example.partymaker.utils.Constants.GROUP_KEY;
+import static com.example.partymaker.utils.Constants.GROUP_LOCATION;
+import static com.example.partymaker.utils.Constants.GROUP_MONTHS;
+import static com.example.partymaker.utils.Constants.GROUP_NAME;
+import static com.example.partymaker.utils.Constants.GROUP_PRICE;
+import static com.example.partymaker.utils.Constants.GROUP_TYPE;
+import static com.example.partymaker.utils.Constants.GROUP_YEARS;
+import static com.example.partymaker.utils.Constants.MESSAGE_KEYS;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
@@ -24,6 +24,7 @@ import android.os.Bundle;
 import android.view.MotionEvent;
 import android.view.View;
 import android.view.ViewConfiguration;
+
 import java.util.HashMap;
 
 /** Utility methods for working with Intents, ExtrasMetadata, and UI helpers in PartyMaker. */
